refactor: remove legacy DataTables path superseded by volumeleaders-go
Delete internal/datatables/ package, legacy DataTables helpers in
cli/common, five dead Post* methods from client, and the unused
DataTablesResponse model. All CLI commands already use the vlgo client
path exclusively since PR #99.
