+++ internal/cli/common/vldatatables.go
@@ -199,3 +199,26 @@ func MapSlice[Src, Dst any](src []Src, mapFn func(*Src) Dst) []Dst {
 	}
 	return dst
 }
+
+// dataTablePageState tracks pagination offset and page size across sequential
+// DataTables requests.
+type dataTablePageState struct {
+	initialStart int
+	pageSize     int
+	optionsValue DataTableOptions
+}
+
+func newDataTablePageState(opts DataTableOptions, pageSize int) dataTablePageState {
+	opts.Length = pageSize
+	return dataTablePageState{initialStart: opts.Start, pageSize: pageSize, optionsValue: opts}
+}
+
+func (state *dataTablePageState) options() DataTableOptions { return state.optionsValue }
+
+func (state *dataTablePageState) fetchedAllRecords(fetched, recordsFiltered int) bool {
+	return recordsFiltered > 0 && state.initialStart+fetched >= recordsFiltered
+}
+
+func (state *dataTablePageState) isShortPage(length int) bool { return length < state.pageSize }
+
+func (state *dataTablePageState) advance(length int) { state.optionsValue.Start += length }
