+++ internal/client/client_test.go
@@ -1,14 +1,9 @@
 package client
 
 import (
-	"compress/gzip"
-	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
-	"strings"
 	"testing"
 	"time"
 
@@ -18,253 +13,6 @@ import (
 func newTestClient(baseURL string) *Client {
 	httpClient := &http.Client{Timeout: 5 * time.Second}
 	return NewForTesting(httpClient, baseURL)
-}
-
-func TestPostDataTables(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		statusCode int
-		body       string
-		gzipBody   bool
-		want       []string
-		wantErr    string
-	}{
-		{
-			name:       "valid response",
-			statusCode: http.StatusOK,
-			body:       `{"draw":1,"recordsTotal":1,"recordsFiltered":1,"data":["AAPL","NVDA"]}`,
-			want:       []string{"AAPL", "NVDA"},
-		},
-		{
-			name:       "empty data field",
-			statusCode: http.StatusOK,
-			body:       `{"draw":1}`,
-			want:       []string{},
-		},
-		{
-			name:       "null data field",
-			statusCode: http.StatusOK,
-			body:       `{"draw":1,"data":null}`,
-			want:       []string{},
-		},
-		{
-			name:       "non 200 status",
-			statusCode: http.StatusBadGateway,
-			body:       `upstream failed`,
-			wantErr:    "status 502: upstream failed",
-		},
-		{
-			name:       "gzip response",
-			statusCode: http.StatusOK,
-			body:       `{"draw":1,"recordsTotal":1,"recordsFiltered":1,"data":["MSFT"]}`,
-			gzipBody:   true,
-			want:       []string{"MSFT"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				if r.Method != http.MethodPost {
-					t.Errorf("expected POST, got %s", r.Method)
-				}
-				if r.URL.Path != "/datatable" {
-					t.Errorf("expected path /datatable, got %s", r.URL.Path)
-				}
-				body, err := io.ReadAll(r.Body)
-				if err != nil {
-					t.Errorf("read request body: %v", err)
-				}
-				if string(body) != "draw=1" {
-					t.Errorf("expected request body draw=1, got %q", string(body))
-				}
-				if tt.gzipBody {
-					w.Header().Set("Content-Encoding", "gzip")
-				}
-				w.WriteHeader(tt.statusCode)
-				writeResponse(t, w, tt.body, tt.gzipBody)
-			}))
-			t.Cleanup(server.Close)
-
-			client := newTestClient(server.URL)
-			var got []string
-			err := client.PostDataTables(t.Context(), "/datatable", "draw=1", &got)
-			assertErrorContains(t, err, tt.wantErr)
-			if tt.wantErr != "" {
-				return
-			}
-			if fmt.Sprint(got) != fmt.Sprint(tt.want) {
-				t.Errorf("expected %v, got %v", tt.want, got)
-			}
-		})
-	}
-}
-
-func TestPostJSON(t *testing.T) {
-	t.Parallel()
-
-	t.Run("valid response", func(t *testing.T) {
-		t.Parallel()
-
-		server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if got := r.Header.Get("Content-Type"); got != "application/json" {
-				t.Errorf("Content-Type: expected application/json, got %q", got)
-			}
-			var payload map[string]string
-			if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
-				t.Errorf("decode request payload: %v", err)
-			}
-			if payload["symbol"] != "AAPL" {
-				t.Errorf("expected symbol AAPL, got %q", payload["symbol"])
-			}
-			fmt.Fprint(w, `{"ok":true}`)
-		}))
-		t.Cleanup(server.Close)
-
-		client := newTestClient(server.URL)
-		var got struct {
-			OK bool `json:"ok"`
-		}
-		if err := client.PostJSON(t.Context(), "/json", map[string]string{"symbol": "AAPL"}, &got); err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if !got.OK {
-			t.Errorf("expected OK response")
-		}
-	})
-
-	t.Run("marshal error", func(t *testing.T) {
-		t.Parallel()
-
-		client := newTestClient("http://example.test")
-		err := client.PostJSON(t.Context(), "/json", map[string]any{"bad": make(chan int)}, &struct{}{})
-		assertErrorContains(t, err, "marshal JSON request")
-	})
-
-	t.Run("non 200 status", func(t *testing.T) {
-		t.Parallel()
-
-		server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-			http.Error(w, "nope", http.StatusInternalServerError)
-		}))
-		t.Cleanup(server.Close)
-
-		client := newTestClient(server.URL)
-		err := client.PostJSON(t.Context(), "/json", map[string]string{"ok": "true"}, &struct{}{})
-		assertErrorContains(t, err, "status 500")
-	})
-}
-
-func TestPostForm(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		statusCode int
-		body       string
-		result     any
-		wantName   string
-		wantErr    string
-	}{
-		{
-			name:       "valid response",
-			statusCode: http.StatusOK,
-			body:       `{"name":"AAPL"}`,
-			result: &struct {
-				Name string `json:"name"`
-			}{},
-			wantName: "AAPL",
-		},
-		{
-			name:       "nil result skips decode",
-			statusCode: http.StatusOK,
-			body:       `not json`,
-		},
-		{
-			name:       "non 200 status",
-			statusCode: http.StatusBadRequest,
-			body:       `bad form`,
-			result:     &struct{}{},
-			wantErr:    "status 400: bad form",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				body, err := io.ReadAll(r.Body)
-				if err != nil {
-					t.Errorf("read request body: %v", err)
-				}
-				if string(body) != "ticker=AAPL" {
-					t.Errorf("expected form body ticker=AAPL, got %q", string(body))
-				}
-				w.WriteHeader(tt.statusCode)
-				fmt.Fprint(w, tt.body)
-			}))
-			t.Cleanup(server.Close)
-
-			client := newTestClient(server.URL)
-			err := client.PostForm(t.Context(), "/form", url.Values{"ticker": {"AAPL"}}, tt.result)
-			assertErrorContains(t, err, tt.wantErr)
-			if tt.wantName != "" {
-				got := tt.result.(*struct {
-					Name string `json:"name"`
-				})
-				if got.Name != tt.wantName {
-					t.Errorf("expected name %q, got %q", tt.wantName, got.Name)
-				}
-			}
-		})
-	}
-}
-
-func TestPostMultipart(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		statusCode int
-		wantErr    string
-	}{
-		{name: "successful post", statusCode: http.StatusOK},
-		{name: "redirect success", statusCode: http.StatusFound},
-		{name: "error status", statusCode: http.StatusBadRequest, wantErr: "status 400: multipart failed"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				if !strings.HasPrefix(r.Header.Get("Content-Type"), "multipart/form-data; boundary=") {
-					t.Errorf("expected multipart content type, got %q", r.Header.Get("Content-Type"))
-				}
-				if err := r.ParseMultipartForm(1024); err != nil {
-					t.Errorf("parse multipart form: %v", err)
-				}
-				if got := r.FormValue("ticker"); got != "AAPL" {
-					t.Errorf("expected ticker AAPL, got %q", got)
-				}
-				if tt.statusCode == http.StatusFound {
-					w.Header().Set("Location", "/next")
-				}
-				w.WriteHeader(tt.statusCode)
-				fmt.Fprint(w, "multipart failed")
-			}))
-			t.Cleanup(server.Close)
-
-			client := newTestClient(server.URL)
-			err := client.PostMultipart(t.Context(), "/multipart", map[string]string{"ticker": "AAPL"})
-			assertErrorContains(t, err, tt.wantErr)
-		})
-	}
 }
 
 func TestMiddlewareHeaders(t *testing.T) {
@@ -334,21 +82,6 @@ func TestMiddlewareCookies(t *testing.T) {
 	}
 }
 
-func writeResponse(t *testing.T, w http.ResponseWriter, body string, gzipBody bool) {
-	t.Helper()
-
-	if !gzipBody {
-		fmt.Fprint(w, body)
-		return
-	}
-	w.Header().Set("Content-Encoding", "gzip")
-	gz := gzip.NewWriter(w)
-	defer gz.Close()
-	if _, err := gz.Write([]byte(body)); err != nil {
-		t.Errorf("write gzip response: %v", err)
-	}
-}
-
 func TestNewForTesting(t *testing.T) {
 	t.Parallel()
 
@@ -365,87 +98,6 @@ func TestNewForTesting(t *testing.T) {
 	}
 	if c.noRedirectClient == nil {
 		t.Errorf("expected no-redirect Resty client to be populated")
-	}
-}
-
-func TestPostDataTablesDataDecodeError(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		fmt.Fprint(w, `{"draw":1,"data":"not an array"}`)
-	}))
-	t.Cleanup(server.Close)
-
-	c := newTestClient(server.URL)
-	var got []string
-	err := c.PostDataTables(t.Context(), "/test", "draw=1", &got)
-	assertErrorContains(t, err, "decode DataTables data")
-}
-
-func TestPostJSONDecodeError(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		fmt.Fprint(w, `not json at all`)
-	}))
-	t.Cleanup(server.Close)
-
-	c := newTestClient(server.URL)
-	var got struct{ Name string }
-	err := c.PostJSON(t.Context(), "/test", struct{}{}, &got)
-	assertErrorContains(t, err, "decode JSON response")
-}
-
-func TestPostFormDecodeError(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		fmt.Fprint(w, `not json`)
-	}))
-	t.Cleanup(server.Close)
-
-	c := newTestClient(server.URL)
-	var got struct{ Name string }
-	err := c.PostForm(t.Context(), "/test", url.Values{"key": {"val"}}, &got)
-	assertErrorContains(t, err, "decode form response")
-}
-
-func TestDoRequestConnectionError(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {}))
-	server.Close()
-
-	c := newTestClient(server.URL)
-	err := c.PostJSON(t.Context(), "/test", struct{}{}, &struct{}{})
-	assertErrorContains(t, err, "post JSON request")
-}
-
-func TestPostMultipartConnectionError(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {}))
-	server.Close()
-
-	c := newTestClient(server.URL)
-	err := c.PostMultipart(t.Context(), "/test", map[string]string{"key": "val"})
-	assertErrorContains(t, err, "post multipart request")
-}
-
-func assertErrorContains(t *testing.T, err error, want string) {
-	t.Helper()
-
-	if want == "" {
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		return
-	}
-	if err == nil {
-		t.Fatalf("expected error containing %q", want)
-	}
-	if !strings.Contains(err.Error(), want) {
-		t.Fatalf("expected error containing %q, got %v", want, err)
 	}
 }
